Move calls back to workspace controller
